Many minor bug fixes and cosmetic changes
Made font adjust better.
Added appearing animation to timetable.
Removed the modified bootstrap css file.
Removed bwip-js and replaced it with jsbarcode, leading to a 700kb size reduction.
Moved the location of main.js when the page is loaded to fix a bug where it would flash when loading dark mode.
Fixed a bug where a filter with no results would break the announcements page.
Also fixed an error where trying to fix an installed version of the app would actually break it. (Sort of.) Very ironic.