Many minor cosmetic and functional updates
Made it so you can't drag the background animation on the settings page.
Removed unused files.
Changed the webp logos to match with the svg one.
Replaced the <h3> tags in the callback and login pages with <p> tags, but with a larger font size.
Also made the images on the callback and login pages non-draggable.
Finally, made it so that if there was no code provided to the callback page, it wouldn't send a request to the servers for a token.